Restructured and decluttered
