+++ src/3d-project/core/animation-loop.ts
