+++ src/3d-project/core/types.ts
