+++ src/views/HomeView.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
-import SceneContainer from "../components/SceneContainer.vue";
+import ThreeController from "@/components/ThreeController.vue";
 </script>
 
 <template>
-  <SceneContainer />
+  <ThreeController />
 </template>
